[dagster-fivetran] Add state storage configuration option for FivetranAccountComponent (#32423)
## Summary & Motivation

This adds a transparent way for users to opt into the more principled state handling APIs provided by the new StateBackedComponents framework.

All StateBackedComponents now have a `defs_state_config` property, which will almost always be populated via a `defs_state` resolvable field on the model (although could in theory be populated some other way for highly-specialized use cases).

In order for this to work, I did end up needing to abandon the use of the `@fivetran_assets` decorator within the body of the component. There's just no realistic way for this to work now that we need to have fine-grained control over how the specs are produced. To be more specific, the decorator currently always invokes `workspace.load_asset_specs()` in order to get the specific set of AssetSpecs that should map to a given connector. Buried within that method call, we end up invoking FivetranWorkspaceDefsLoader.build_defs(), which is highly opinionated about the key that is used within the reconstruction metadata. In order for this to work now that StateBackedComponents are taking responsibility for the state handling process, we would need to thread through the state directly into workspace.load_asset_specs(). But even that doesn't work great -- beyond polluting a public API, we'd need to handle both cases of:

- state is passed through externally (state-backed components)
- state is not passed through (current-day)

To do this, it's not enough to just have an `Optional[TState]` argument on `build_defs()`, because we need to handle the case where state backed components are managing state for the user but there's not yet any state available. If you're using REMOTE / LOCAL storage, we don't want to then hit the fivetran API in these cases, so we need something more than a `None` value to distinguish between the cases.

I think the simplicity in this implementation is well worth the mild risk of divergence (i.e. if we decide to change anything about the `@fivetran_assets` decorator, that will not automatically be reflected in the component). 


## How I Tested These Changes

## Changelog

NOCHANGELOG