[dagster-dbt] Make DbtProjectComponent a StateBackedComponent (#32486)
## Summary & Motivation

This turns the DbtProjectComponent into a state-backed component to accurately represent the fact that it depends on the state of the `manifest.json` file. This generalization means that running `dg utils refresh-defs-state` within the context of a docker build step will manage building all of the state across all integrations. no more bespoke CLIs specifically for dbt or other integrations

This also (perhaps more importantly) opens up the door for managing dbt projects that don't live in the same repo as the dagster project.

For now, only `LOCAL_FILESYSTEM` is supported here, but we could imagine a world in which `VERSIONED_STATE_STORAGE` is supported as well.

## How I Tested These Changes

## Changelog

> Insert changelog entry or delete this section.