[sbc] Keep track of which defs keys were actually accessed during the load process (#32504)
## Summary & Motivation

This is more for vanity than anything else, but consider the following situation:

1. code server starts up with no explicit DefsStateInfo passed in (this is the default in non-plus deployments)
2. DefinitionsLoadContext sees this, and so automatically pulls in the latest versions from state storage
3. when we're serializing metadata about our workspace, we pull this off of the DefinitionsLoadContext.

For that final step, this means that we're including the latest version of ALL defs keys that the instance knows about, which can include keys for old state backed components that are no longer relevant. This filters out the list that we serialize to just ones that were accessed during the load process, ensuring that we don't show old / confusing stuff.

This is similar in nature to we handle the opposite case, where we attempt to laod a state backed component that doesn't have any corresponding state in storage. In this case, we create an entry in the info_mapping dictionary with a null info value to indicate that we know that this key desires state

## How I Tested These Changes

## Changelog

NOCHANGELOG