[sbc] Move state management under project root (#32523)
## Summary & Motivation

For local filesystem state storage, it's desirable to have the local state stored in a path that's relative to the project root instead of a user data dir, specifically for PEX-deployed projects which require data to be in the python package files. Instead of making that a bespoke aspect of the deploy process, we can just store this in an accessible place from the start.

## How I Tested These Changes

## Changelog

NOCHANGELOG