[sbc] Sanitize path key (#32531)
## Summary & Motivation

This prevents issues that can arise if the key contains weird characters. The worst offender here is a key of the form `FooComponent[https://foo.com]`, in which that gets converted to the file path `path/to/FooComponent[https:/foo.com]` (note that `FooComponent[https:` is a directory).

This vaguely kinda doesn't cause issues, but is really ugly and it's easier to change it now than later.

## How I Tested These Changes

## Changelog

NOCHANGELOG