[ui] Replace react-dates with react-day-picker (#32525)
## Summary & Motivation

`react-dates` hasn't had a version published in 6 years, and it depends on `moment`, which years ago went into maintenance mode.

Replace this with `react-day-picker`, which uses `date-fns`.

Right now it's only used in `useTimeRangeFilter`, but we're going to use it in Observe as well.

![Screenshot 2025-10-15 at 09.57.47.png](https://app.graphite.dev/user-attachments/assets/6e813642-6a4e-4519-b302-beefb1b956f1.png)

## How I Tested These Changes

Storybook examples, jest. View runs page in app, verify that I can set a custom time range in the filter.

## Changelog

[ui] Update custom time range component.