[sbc] Don't error if DefsStateStorage not available for local / code-server storage types (#32559)
## Summary & Motivation

As title. This is unnecessary for loading these types of state so we
shouldn't aggressively error if the instance is not available. We should
update the default behavior in k8s deployments to include access to the
instance regardless of this issue (working on it!) but for now this is a
simple and generally useful fix.

## How I Tested These Changes

## Changelog

Fixed a bug that caused errors when using the `DbtProjectComponent`,
`FivetranAccountComponent`, and similar state-based components in k8s
deployments due to a missing `StateStorage` object in context.