fix lazy definition validation
Summary:
We were previously validating that the key matched the name only if it was a lambda. Now validate it no matter what.

BK

Fixed an issue where definining a repository using a dictionary of job definitions with a key that did not match the name of the job would work when running dagster locally but not when using Dagster+.

> Insert changelog entry or delete this section.