fix lazy definition validation (#32561)
## Summary & Motivation
We were previously validating that the key matched the name only if it
was a lambda. Now validate it no matter what.

## How I Tested These Changes
BK

## Changelog
Fixed an issue where definining a repository using a dictionary of job
definitions with a key that did not match the name of the job would work
when running dagster locally but not when using Dagster+.