[sbc] Update helm chart to allow instance configuration for user code pods (#32249)
## Summary & Motivation

We need user code pods to have access to the instance in order for state backed components to work. This is generally possible as most users deploy their infra / user code in the same namespace (and so will have access to the instance configmap), but we also need to handle the case where this is not the case.

I originally wanted to make this mounting process optional but unfortunately if you set a `subpath` for a volume mount and that subpath does not exist, k8s (insanely) takes that to mean that it should mount the entire volume at that path instead. This means that `dagster.yaml` becomes an empty directory which messes up instance loading.

So for now, this will be opt-in, and then once 1.12.0 releases, we can make this opt-out.

Finally, I've updated the server startup process to always enter the instance context if possible (i.e. there is a DAGSTER_HOME set or an instance_ref was passed in). In OSS deployments, `defs_state_info` will not be explicitly passed in at server startup time

## How I Tested These Changes

## Changelog

NOCHANGELOG