[dagster-dbt] Add get_resource_props method (#32598)
## Summary & Motivation

The get_asset_spec pattern is generally way better for users than the old-style pattern where we would have a bunch of different `get_x` methods, but one consequence is that the method signature for get_asset_spec in dagster-dbt only has higher level objects, rather than the simpler resource_props dictionary. In order to get that, it's not actually trivial (not just indexing into the manifest dictioanry), and so rather than  forcing users to do that themselves or exporting the get_node() method to the top-level of dagster-dbt, it felt cleaner to make this a method  on the relevant classes.

## How I Tested These Changes

## Changelog

> Insert changelog entry or delete this section.