[record] have pickle bypass runtime checks / custom logic (#32610)
We only use `pickle` to transport (as far as I understand it)
information between processes of the same version, so this change makes
it so we inflate the underlying namedtuple directly bypassing runtime
checks and custom logic. This improves the efficiency of using pickle in
this same version scenario.

## How I Tested These Changes

existing record pickle tests