[da] Update `evaluation_id` threading (#32628)
## Summary & Motivation

We no longer use the `evaluation_id` that's set on the AssetDaemonCursor for anything, instead opting for getting the evaluation id from the database when we generate a new tick. This means that code that was trying to access this during AutomationCondition evaluation would get an incorrect value (this was just the SinceCondition metadata, so it didn't impact anything functional).

## How I Tested These Changes

## Changelog

NOCHANGELOG