FE-994: Split DateRangeDialog out into a separate component (#32651)
## Summary & Motivation

This PR splits the dialog version of the date range component out from
the time filter into a separate `DateRangeDialog`.

I thought about just copy pasting this, but there are a few things that
are tricky that are done well here: It lazy loads the date picker, which
required a separate shim file, and it coerces to end-of-day in the right
timezone. I think that it's worth discouraging more copies of this and
trying to use this one more generally.

This takes no initial value because so far both use cases are for filter
entry, and i think it's better to choose from scratch each time.

## Testing

I verified the "Created at" filter on the Jobs page still works as
expected

---------

Co-authored-by: bengotow <bgotow@elementl.com>