fix: deltalake string partition dimension str_values (#19888)
## Summary & Motivation

A missing flag causes an error when loading an asset with deltalake io
manager when it is partitionned with a string-valued column. This case
was not covered in the original unit tests.

Without this flag, the following error appears : 

```
pyarrow.lib.ArrowNotImplementedError: Function 'equal' has no kernel matching input types (string, list<item: string>)
```

## How I Tested These Changes

Wrote a new unit test