[sbc] Update scaffolded github action to use `dg plus deploy refresh-defs-state` (#32678)
## Summary & Motivation

Updates the scaffold templates to include commands to refresh defs state. In the process of this, I realized our existing docs were way too complicated because there's really not much use case for doing any def refreshing inside your Dockerfile (you can just do it outside and the copy the state inside). This was a relic of an earlier setup where state was not written to a directory that was accessible in this way.

So updated the docs accordingly

## How I Tested These Changes

Manually in internal.

## Changelog

The github actions scaffolded by `dg scaffold github-actions` now include commands to refresh state for `StateBackedComponents`.