use has_permission_for_definition instead of has_permission_for_selector in asset queries
Summary:
This is faster when you already have the remote node loaded, which we do here.

However, one thing we need to verify is how the frontend combines the results of these permissions for different repositories. This is one where you really only want to check it for the materialiable one if it is present I think?