Fix race condition when multiple threads are generating job snapshots/job config at once
Summary:
The old logic here isn't thread-safe - If thread B fetches config while thread A is in the middle of resolving config, the state might be unfinished and return unexpected results. Instead, embrace the memoization and keep the object immutable.

Working on a test, but a script

Fixed a sporadic race condition when loading jobs in the Dagster UI.