Fix race condition when multiple threads are generating job snapshots/job config at once (#32686)
## Summary & Motivation
The old logic here that was added in
https://github.com/dagster-io/dagster/pull/30453 isn't thread-safe - If
thread B fetches config while thread A is in the middle of resolving
config, the state might be unfinished and return unexpected results.
Instead, embrace the memoization and keep the object immutable.

## How I Tested These Changes
Working on a test, but a script that concurrently accesses multiple job
snapshots at once now returns consistent results.

## Changelog
Fixed a sporadic race condition when loading jobs in the Dagster UI.