[ui] Ensure only asset key is provided as arg in Asset permissions query (#32754)
## Summary & Motivation

A GraphQL error currently shows up in OSS dev when viewing an asset detail page, in which `AssetPermissonsQuery` fails because the asset key arg being provided includes a `__typename`.

This means that upstream, a full GraphQL object (including `__typename`) is being passed through, and TypeScript doesn't object because it satisfies the required object properties for the input value.

To resolve this, just use the `path` value for the argument.

## How I Tested These Changes

Local dev, view asset detail page for an SDA. Verify that there are no GraphQL errors.