[dg] refactor `dg scaffold github-actions` and `dg scaffold build-artifacts` into `dg plus deploy configure` (#32723)
## Summary & Motivation

As title. Creates a single entrypoint for getting an existing project ready for deployment.

Also allows all parameterization to be (optionally) done via command line flags to help with automation from external tools / workflows.

Structurally, this introduces a single entrypoint where parameters can be passed in optionally. if parameters are not passed in, we try to pull them from the environment (e.g. reading config files). if we are not able to pull them from the environment, we prompt the user for them.

This setup allows us to present a user-friendly prompt-based workflow, while also enabling tools to just inject the parameters directly.

The existing github-actions and build-artifacts commands can then just invoke this same command for backcompat.

Also sets us up nicely for adding gitlab support (this is factored in a way that we just pass in a git_provider param so we can always add more there).

## How I Tested These Changes

## Changelog

Added a new `dg plus deploy configure` CLI group that generates all the files necessary to get an existing project ready for deployment via Dagster+.