[ui] Add a hierarchical sidebar to the asset catalog (#32775)
## Summary & Motivation
<img width="1728" height="734" alt="image"
src="https://github.com/user-attachments/assets/999801a2-37e3-43e8-ab10-ef7a46026d73"
/>

This PR extracts some reusable bits from existing components to create a
new `HierarchicalSidebar` component that is fairly content-agnostic, and
then uses it in the asset catalog. Our existing asset graph sidebar is
great but is very specifically built for the lineage views and can't
easily be reused.

When you select a node in the sidebar, we update the asset selection to
apply an additional `key:prefix/*` filter or update the existing one.
This uses two utility methods that adjust the selection string. I think
that a more exhaustive version of this might use `SelectionInputParser`,
but I couldn't think of cases that would handle that the `if an OR
clause exists, wrap the selection in parens and add AND key:prefix`
doesn't address.

A nice direction for future work would be mulit-key-prefix-selection,
but it'd require auto-building `(key: A OR key: B) AND ( ...rest... )`
and get pretty unweildly pretty fast.

## How I Tested These Changes

- Manual testing and some new tests covering selection updating and tree
building

## Changelog

[ui] A new sidebar in the Asset catalog (behind a feature flag) provides
a hierarchical view of your asset keys.

## Feature Flagging

This is not currently feature flagged. We discussed doing a Statsig
gate, but I've implemented it as a front-end flag for the upcoming
release since the code lives in OSS.

---------

Co-authored-by: bengotow <bgotow@elementl.com>