Speed up instigatorStateOrError
Summary:
Circumvent the potentially slow repository fetch here. A multi-get would be even better so that we can check for sensors and schedules in parallel (or passing in whether its a schedule or a sensor as a hint) but this should still be a big improvement even for schedules.

Test Plan:
Existing test coverage, dry runs