Fix cursor unique id generation (#32826)
## Summary & Motivation

This makes it so that, for SINCE / NEWLY_TRUE conditions specifically, we can retain memory of their value no matter where they move in the tree.

This is safe for these conditions because they always execute against the entire set of partitions of an asset, regardless of what their candidate subset is. This means that their result value is not actually influenced by anything other than their sub conditions and the key that they are being applied against.

This uses the existing backcompat_node_unique_ids() functionality to ensure that we can still find cursors stored with the old serialization format.

## How I Tested These Changes

## Changelog

Updated the cursoring logic of `AutomationCondition.since()`/`AutomationCondition.newly_true()` to make them retain access to their stored data in a wider range of scenarios where the underlying condition structure is changed.