[bug] Fix issue with applying metadata to sensors (#32810)
## Summary & Motivation

Fixed the added test, unfortunately no easier way of doing this as far as I'm aware.

Core problem is that the old version of this function would always return an instance of SensorDefinition, rather than the appropriate subclass.

Added dagster_internal_init methods to make it a bit less likely that we forget to add a parameter, but pretty painful all throughout.

## How I Tested These Changes

## Changelog

Fixed an issue that would cause `SensorDefinition` subclasses (e.g. `AutomationConditionSensorDefinition`, `RunStatusSensorDefinition`) to be converted to having the wrong `sensor_type` property when produced from a `Component`.