Don't include the value of an environment variable in the error message for an IntSource or BoolSource failing to validate it (#32914)
## Summary & Motivation
Since the reason these are in environment variables is likely becauase they are sensitive, ensure that the error message (including the stack trace/context) does not include the actual value. Makes things marginally more difficult to debug but seems like the right tradeoff in this case.

## How I Tested These Changes
BK

## Changelog
Errors for IntSource and BoolSource config types failing to validate an environment variable value will no longer include the value of the environment variable in the error message.