[da] Add additional forwards-compatibility logic / test (#32928)
## Summary & Motivation

We need to be a bit hacky here to get the precise correct forwards-compat behavior. In essence, we want this code to be able to generate the exact same hash that the code in the upstack PR would generate (in this PR, it does it in the backcompat path, and in the upstream PR it happens in the main path).

The issue is that _get_stable_unique_id calls get_node_unique_id, and the upstream PR changes the result of that for the SinceCondition/NewlyTrueCondition classes. So we add in a little bit of logic here to invoke the right method in these cases. We remove this weird workaround logic in the upstack PR for clarity (as in that PR, both of those methods are the same)

## How I Tested These Changes

## Changelog

> Insert changelog entry or delete this section.