[dg] Update `dg plus deploy refresh-defs-state` command to handle subprocesses (#32851)
## Summary & Motivation

This makes it so `dg plus deploy refresh-defs-state` can handle workspaces in addition to projects. previously we assumed we were executing in the same python environment as the main dg command, but this isn't always the case. Now, we create an InstanceRef that has access to our prod DefsStateStorage. We can then serialize this reference and pass it to the subprocess, which then invokes `dg utils refresh-defs-state` *in the project's python env*. This results in the instance in the main process being mutated, which we can then read from at the end to get the updated defs state info.

## How I Tested These Changes

## Changelog

> Insert changelog entry or delete this section.