Transform JsonMetadataValue before storing it (#32942)
## Summary & Motivation
This fixes an issue where you could pass in a string enum to
JsonMetadataValue and it would pass this validation check, but fail when
you later try to serdes it into the database.

## Changelog
Fixed an issue where passing in JSON serializable enums to
JsonMetadataValue would sometimes result in an error.