[ui] Fix icon display in list items (#32981)
## Summary & Motivation

@braunjj noticed that the insights icons in catalog lists are rendering with their original color, and with weird background behavior.

This is related to the recent Icon refactor, and the fact that ListItem has CSS that explicitly sets `background-color` on icons. I think we can remove the ListItem CSS and require callers to set their icon color themselves. Then, by setting icon color explicitly on these `InsightsIcon` use cases, we'll get the correct colors again.

## How I Tested These Changes

View Catalog, verify that insights icons are rendered with the same color as other icons in the list.

## Changelog

[ui] Fix icon colors in list items.