[ui] TS native preview (#32990)
Add the TypeScript native preview to a few JS projects in `dagster`, along with a `tsgo` script to run it.

https://devblogs.microsoft.com/typescript/progress-on-typescript-7-december-2025/

This will let us start trying it out a bit before everyone switches to use it, using it in builds, etc.

I had to iron out a couple of issues. Will comment inline.

## Test Plan

For each project with `tsgo` added, run `yarn tsgo`. `make tsgo` from `dagster-ui`.