[secret-resource-param] core (#32991)
## Summary & Motivation

Add support for marking resource parameters as secret, allowing them to
be hidden in the UI. This is done by using `pydantic.Field`'s
`json_schema_extra` property, which is intended for this sort of
metadata.

Note that this is an alternative approach to solving the same issue that
https://github.com/dagster-io/dagster/pull/14044 tried to solve in the
past. While using the untyped `json_schema_extra` field is a little
unfortunate (we can't add a parameter directly in the user-facing API
since we use pydantic `Field` rather than our internal `Field`), it does
have the advantage of avoiding any new Dagster concept/entities.

```
from pydantic import Field as PyField

class MyResource(ConfigurableResource):
    foo: str = Field(json_schema_extra={"is_secret": True})
```

Upstack we pipe through GQL and up to the UI.

## How I Tested These Changes

New unit tests.

## Changelog

> Insert changelog entry or delete this section.