[nux] Fix refresh_defs_state handling of subprocesses (#33023)
## Summary & Motivation

While the code before did use the correct python executable, and so all
python dependencies were correctly handled, if the code required
invoking a CLI that was only available in the project environment, this
would fail because we did not update the `PATH` to include this.

This uses the activate_venv utility to update the path -- it also
updates the VIRTUAL_ENV env var which avoids an annoying warning we were
getting before about the project env not matching (even though it did --
we just didn't have the right env var set)

## How I Tested These Changes

## Changelog

Fixed an issue with `dg plus deploy refresh-defs-state` which could
cause errors when refreshing state for components that required CLIs
that were only available in the project environment.