Bump asset kind limit to 10 (#32982)
## Summary & Motivation

Bumps limit on number of kinds that can be set on an asset to 10
(previously 3).

Kinds do start to look a little weird if they get too long (see
screenshot) but this is already an issue with 3, if the text is long
enough.

Resolves #32976


![image.png](https://app.graphite.com/user-attachments/assets/587f4a8a-3e6f-4c49-9891-ab04b5b918f9.png)

Have confirmed with @benpankow that there's no reason for the limit of 3
other than it starting to look odd in the UI.

## How I Tested These Changes

Existing test suite.

## Changelog

Assets may now be annotated with up to 10 kinds (limit was previously
3).