Better handling for backfill runs materializing unpartitioned materializations when they are supposed to be partitioned (and vice versa) (#33035)
## Summary & Motivation
Before these would move the backfill object into an invalid state where the materialized subset was incompatible with the target subset. Now we detect the invalid case during the backfill and fail it.

## How I Tested These Changes
New test cases

## Changelog
Fixed an issue where backfill runs that incorrectly created unpartitioned materializations of a partitioned asset, or partitioned materializations of an unpartitioned asset due to incorrect asset business logic would move the backfill into an invalid state. Now, the backfill will detect this case and fail the backfill.