[ui] Try to fix weird resize loop on asset plot (#32978)
## Summary & Motivation

One of our users is reporting strange behavior with asset plot charts, where the chart automatically resizes itself smaller and smaller.

I'm going to try to fix this by putting a wrapper div with a set height, and remove the `width` and `height` props from the `Line` component.

I'm not able to repro the issue myself, but this approach seems worth a try.

## How I Tested These Changes

View asset plots, verify that they render properly.