copy material-icons icon exactly (#33055)
## Summary & Motivation

Make the new icons appear correctly

## How I Tested These Changes

Ran locally, confirmed icon appearance in the UI.

## Notes

I tried to change some of the metadata in the svg opening tag to align
with the other svgs in the repo more closely, but this broke the svg. I
didn't notice because I was doing the development with another svg until
this was merged. This PR reverts the changes to exactly match what was
in the
https://github.com/marella/material-design-icons/blob/main/svg/outlined
(thumb_up|thumb_down).svg files.