Fix handling of pythonic config with PEP 649 (#33051)
## Summary & Motivation

Pretty gross, but this was the best I could figure out. PEP 649 pretty
signficantly changes how annotations are handled on classes in ways that
break our existing utilities. In particular, because annotations are
lazily evaluated by default, some of our logic which attempts to
intercept the annotations to check for things like
`ResourceDependency[T]` things would no longer work because we were
explicitly analyzing the `__annotations__` value.

This PR executes the `__annotate_func__` to make sure we have the
fully-instantiated types

## How I Tested These Changes

Ran relevant tests in an env with python 3.14 installed -- failed
before, pass now. The next pr upstack adds this testing to BK, just
split it up here to isolate the spicy changes.

## Changelog