[asset-deps] Refactor ResolvedAssetDependencies (#33018)
## Summary & Motivation

I can't recall the exact series of events that lead this to being its own separate class, but in today's world this is just unnecessarily roundabout. We're only using this functionality in a single place, so just having the function directly produce a list of resolved asset dependencies prevents a whole slew of roundabout manipulations.

This is not intended to change any functionality, just makes it easier to do manipulations upstack.

## How I Tested These Changes

## Changelog

> Insert changelog entry or delete this section.