[ui components] Button CSS (#32963)
## Summary & Motivation

Refactor `Button` in `ui-components` to us CSS modules instead of
styled-components.

Additionally, eliminate the intermediate `StyledButton` and `BaseButton`
components, which expect prop values to set colors and styles, but which
we don't really use in practice and just add layers of complexity to an
otherwise simple component.

## How I Tested These Changes

Storybook for examples.

Click around the app in OSS and Plus, verify that buttons render and
behave correctly, including joined buttons, disabled buttons, focus
states, etc.