[asset-deps] Include AssetDep metadata in stub assets (#33026)
## Summary & Motivation

As title. This makes it so the external stub assets that we automatically generate contain metadata that was associated with the deps that reference them. For now we enforce that all dependencies that reference a single key must have the same exact metadata, or no metadata at all, but we may relax this in the future

## How I Tested These Changes

## Changelog

> Insert changelog entry or delete this section.