[dagster-dbt] Be more resilient to missing timing data from dbt cloud (#33045)
## Summary & Motivation

It's possible for dbt cloud to not return any timing data in some circumstances. The main purpose of us tracking this is to be able to emit materialization events in the proper order. This gives a fallback of using the time we received the event as the completed_at time, and swaps the tuple that we're using to sort our materialization events to prioritize the topological order

## How I Tested These Changes

## Changelog

[dagster-dbt] Fixed issue that could cause errors when emitting events for a dbt Cloud job run