[ui] Show appropriate reason on tooltip when disabling sensor switch (#33063)
## Summary & Motivation

We currently retrieve the permissions for starting/stopping a sensor for
rendering the sensor switch, but we don't show a good reason to
represent the disabled state when permissions are false.

This reason value doesn't appear to be available via GraphQL, but I
think it should be okay to use the permissions for the code location, so
use that here.

![Screenshot 2025-12-16 at
20.28.00.png](https://app.graphite.com/user-attachments/assets/82b8070e-4294-406e-ab20-ed20341df9ac.png)

## How I Tested These Changes

View branch deployment, go to an automation. Hover on the switch, verify
that I've got a useful tooltip.

View prod deployment, verify that enabled sensor switches render without
tooltips.

## Changelog

[ui] Improve text for the tooltip on sensor switches when they are
disabled for permissions reasons.